requirements.txt
amqp==5.1.1
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.6
Django==4.1
django-ckeditor-5==0.2.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-recaptcha==3.0.0
django-resized==1.0.2
django-taggit==3.0.0
entrypoints==0.4
executing==1.1.1
filelock==3.8.0
glob2==0.7
ipykernel==6.16.1
ipython==8.5.0
jedi==0.18.1
joblib==1.2.0
jupyter_client==7.4.3
jupyter_core==4.11.2
kombu==5.2.4
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
nltk==3.7
numpy==1.23.3
packaging==21.3
pandas==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
psutil==5.9.3
psycopg2==2.9.4
pure-eval==0.2.2
py-vncorenlp==0.1.3
Pygments==2.13.0
pyjnius==1.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
pywin32==304
pyzmq==24.0.1
redis==4.3.4
regex==2022.9.13
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.1
textblob==0.17.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
tzdata==2022.2
vine==5.0.0
virtualenv==20.16.5
wcwidth==0.2.5
wrapt==1.14.1
